Split StandardErrorResolver into its own file
